requirements.txt
streamlit==1.28.0
pandas==2.0.3
pillow==10.0.1
numpy==1.25.2
